week-3/swarm-app-with-db/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
